+++ src/TALib.NETCore.HighPerf/Candlestick/BeltHold.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode BeltHold(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -38,9 +38,9 @@ namespace TALib.NETCore.HighPerf
                 return RetCode.Success;
             }
 
-            double bodyLongPeriodTotal = default;
+            decimal bodyLongPeriodTotal = default;
             int bodyLongTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyLong);
-            double shadowVeryShortPeriodTotal = default;
+            decimal shadowVeryShortPeriodTotal = default;
             int shadowVeryShortTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowVeryShort);
             int i = bodyLongTrailingIdx;
             while (i < startIdx)
