+++ src/TALib.NETCore.HighPerf/Candlestick/CounterAttack.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode CounterAttack(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -39,7 +39,7 @@ namespace TALib.NETCore.HighPerf
             }
 
             var bodyLongPeriodTotal = BufferHelpers.New(2);
-            double equalPeriodTotal = default;
+            decimal equalPeriodTotal = default;
             int equalTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Equal);
             int bodyLongTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyLong);
             int i = equalTrailingIdx;
