+++ src/TALib.NETCore.HighPerf/Candlestick/GapSideSideWhite.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode GapSideSideWhite(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -38,8 +38,8 @@ namespace TALib.NETCore.HighPerf
                 return RetCode.Success;
             }
 
-            double nearPeriodTotal = default;
-            double equalPeriodTotal = default;
+            decimal nearPeriodTotal = default;
+            decimal equalPeriodTotal = default;
             int nearTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Near);
             int equalTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Equal);
             int i = nearTrailingIdx;
