+++ src/TALib.NETCore.HighPerf/Candlestick/LongLine.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode LongLine(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -38,9 +38,9 @@ namespace TALib.NETCore.HighPerf
                 return RetCode.Success;
             }
 
-            double bodyPeriodTotal = default;
+            decimal bodyPeriodTotal = default;
             int bodyTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyLong);
-            double shadowPeriodTotal = default;
+            decimal shadowPeriodTotal = default;
             int shadowTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowShort);
             int i = bodyTrailingIdx;
             while (i < startIdx)
