+++ src/TALib.NETCore.HighPerf/Candlestick/RickshawMan.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode RickshawMan(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -38,11 +38,11 @@ namespace TALib.NETCore.HighPerf
                 return RetCode.Success;
             }
 
-            double bodyDojiPeriodTotal = default;
+            decimal bodyDojiPeriodTotal = default;
             int bodyDojiTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyDoji);
-            double shadowLongPeriodTotal = default;
+            decimal shadowLongPeriodTotal = default;
             int shadowLongTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowLong);
-            double nearPeriodTotal = default;
+            decimal nearPeriodTotal = default;
             int nearTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Near);
             int i = bodyDojiTrailingIdx;
             while (i < startIdx)
