+++ src/TALib.NETCore.HighPerf/Candlestick/ThreeWhiteSoldiers.cs
@@ -5,10 +5,10 @@ namespace TALib.NETCore.HighPerf
     public static partial class Lib
     {
         public static RetCode ThreeWhiteSoldiers(
-            ref Span<double> inOpen,
-            ref Span<double> inHigh,
-            ref Span<double> inLow,
-            ref Span<double> inClose,
+            ref Span<decimal> inOpen,
+            ref Span<decimal> inHigh,
+            ref Span<decimal> inLow,
+            ref Span<decimal> inClose,
             int startIdx,
             int endIdx,
             int[] outInteger,
@@ -44,7 +44,7 @@ namespace TALib.NETCore.HighPerf
             int shadowVeryShortTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.ShadowVeryShort);
             int nearTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Near);
             int farTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.Far);
-            double bodyShortPeriodTotal = default;
+            decimal bodyShortPeriodTotal = default;
             int bodyShortTrailingIdx = startIdx - CandleAvgPeriod(CandleSettingType.BodyShort);
 
             int i = shadowVeryShortTrailingIdx;
